+++ src/flows/user-program-management/types.ts
@@ -3,6 +3,12 @@ export interface AddProgramParams {
   programConfig: string
   verifyingKey?: string 
 }
+
 export interface ViewProgramsParams {
+  verifyingKey: string
+}
+
+export interface RemoveProgramParams {
+  programPointer: string
   verifyingKey: string
 }