+++ tests/user-program-management.test.ts
@@ -1,13 +1,14 @@
 import test from 'tape'
 import { readFileSync } from 'node:fs'
-import { charlieSeed, charlieStashSeed, setupTest } from './testing-utils'
+import { promiseRunner, charlieStashSeed, setupTest } from './testing-utils'
 import { AddProgramParams } from 'src/flows/user-program-management/types'
 import { addProgram } from 'src/flows/user-program-management/add'
 import { viewPrograms } from 'src/flows/user-program-management/view'
+import { removeProgram } from 'src/flows/user-program-management/remove'
 
 const networkType = 'two-nodes'
 
-test('User Program Management::Add Programs', async t => {
+test('User Program Management', async t => {
   const { run, entropy } = await setupTest(t, { seed: charlieStashSeed, networkType })
   await run('charlie stash register', entropy.register())
   const noopProgram: any = readFileSync(
@@ -23,21 +24,33 @@ test('User Program Management::Add Programs', async t => {
     programConfig: '',
   }
 
-  const programsBeforeAdd = await run('get programs initial', entropy.programs.get(entropy.programs.verifyingKey))
-  t.equal(programsBeforeAdd.length, 1, 'charlie has 1 program')
-  await run('adding program', addProgram(entropy, noopProgramInstance))
-  const programsAfterAdd = await run('get programs after add', entropy.programs.get(entropy.programs.verifyingKey))
-  t.equal(programsAfterAdd.length, 2, 'charlie has 2 programs')
+  t.test('Add Program', async ap => {
+    const runAp = promiseRunner(ap)
 
-  t.end()
-})
+    const programsBeforeAdd = await runAp('get programs initial', entropy.programs.get(entropy.programs.verifyingKey))
+    ap.equal(programsBeforeAdd.length, 1, 'charlie has 1 program')
+    await runAp('adding program', addProgram(entropy, noopProgramInstance))
+    const programsAfterAdd = await runAp('get programs after add', entropy.programs.get(entropy.programs.verifyingKey))
+    ap.equal(programsAfterAdd.length, 2, 'charlie has 2 programs')
+    ap.end()
+  })
 
-test('User Program Management::View Programs', async t => {
-  const { run, entropy } = await setupTest(t, { seed: charlieSeed, networkType })
+  t.test('Remove Program', async rp => {
+    const runRp = promiseRunner(rp)
+    const programsBeforeRemove = await runRp('get programs initial', entropy.programs.get(entropy.programs.verifyingKey))
+  
+    rp.equal(programsBeforeRemove.length, 2, 'charlie has 2 programs')
+    await runRp('removing noop program', removeProgram(entropy, { programPointer: newPointer, verifyingKey: entropy.programs.verifyingKey }))
+    const programsAfterRemove = await runRp('get programs initial', entropy.programs.get(entropy.programs.verifyingKey))
+    rp.equal(programsAfterRemove.length, 1, 'charlie has 1 less program')
+    rp.end()
+  })
 
-  await run('charlie register', entropy.register())
-  const programs = await run('get charlie programs', viewPrograms(entropy, { verifyingKey: entropy.programs.verifyingKey }))
+  t.test('View Program', async vp => {
+    const runVp = promiseRunner(vp)
+    const programs = await runVp('get charlie programs', viewPrograms(entropy, { verifyingKey: entropy.programs.verifyingKey }))
 
-  t.equal(programs.length, 1, 'charlie has 1 program')
-  t.end()
+    vp.equal(programs.length, 1, 'charlie has 1 program')
+    vp.end()
+  })
 })